requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ultralytics>=8.0.0
pillow>=10.0.0
pyyaml>=6.0
numpy>=1.24.0
opencv-python>=4.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
click>=8.1.0
python-dotenv>=1.0.0
requests>=2.31.0
huggingface-hub>=0.17.0
colorama>=0.4.6